+++ egroot/src/internal/equals.c
@@ -1,10 +1,11 @@
 #include <internal/types.h>
 
-bool equals(string s1, string s2) {
+bool
+equals(string s1, string s2) {
 	if (s1.len != s2.len) {
 		return false;
 	}
-	int i;
+	int_ i;
 	for (i = 0; i < s1.len; ++i) {
 		if (s1.str[i] != s2.str[i]) {
 			return false;
